applicant registration

+++ JobPortalSharp.Data/Entities/Applicant.cs
@@ -24,6 +24,6 @@ namespace JobPortalSharp.Data
         public string Address { get; set; }
         public string MobileNumber { get; set; }
         public string PhoneNumber { get; set; }
-        public float ExpectedSalary { get; set; }
+        public decimal ExpectedSalary { get; set; }
     }
 }
